fix: change type for income from double to BigDecimal

+++ src/main/java/dev/otthon/loan/domain/Customer.java
@@ -1,6 +1,7 @@
 package dev.otthon.loan.domain;
 
 import lombok.AllArgsConstructor;
+import java.math.BigDecimal;
 
 @AllArgsConstructor
 public class Customer {
@@ -8,7 +9,6 @@ public class Customer {
     private Integer age;
     private String cpf;
     private String name;
-    private Double income;
+    private BigDecimal income;
     private String location;
-
 }
